Add pkcs11-module-default-slot-id cfg for default slot ID
There are use cases when one wants to select the default slot id
(especially when the first slot ID is not convenient for this purpose)
which is used for example for storing public keys.

As part of this change, the test framework was extended to allow
testing of multi slot scenarios.

Signed-off-by: Jakub Zelenka <jakub.openssl@gmail.com>